tests: fix relation types
